Split file, fix exports
